rename: domain 패키지 밑으로 bakery 패키지 이동 (#15)
