+++ breaditnow-owner-api/src/main/java/com/breaditnow/owner/domain/bakery/controller/req/BakeryUpdateRequest.java
@@ -1,6 +1,6 @@
-package com.breaditnow.owner.bakery.controller.req;
+package com.breaditnow.owner.domain.bakery.controller.req;
 
-public record BakeryCreateRequest(
+public record BakeryUpdateRequest(
 	String name,
 	String addressCode,
 	String addressDescription,
