+++ breaditnow-owner-api/src/main/java/com/breaditnow/owner/domain/bakery/controller/req/OperatingStatusRequest.java
@@ -1,4 +1,4 @@
-package com.breaditnow.owner.bakery.controller.req;
+package com.breaditnow.owner.domain.bakery.controller.req;
 
 public record OperatingStatusRequest(String operatingStatus) {
 }
