+++ breaditnow-owner-api/src/test/java/com/breaditnow/owner/bakery/service/BakeryServiceTest.java
@@ -17,9 +17,10 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.breaditnow.domain.domain.owner.entity.Owner;
 import com.breaditnow.domain.domain.owner.repository.OwnerRepository;
-import com.breaditnow.owner.bakery.controller.req.BakeryCreateRequest;
-import com.breaditnow.owner.bakery.controller.req.BakeryUpdateRequest;
-import com.breaditnow.owner.bakery.controller.res.BakeryResponse;
+import com.breaditnow.owner.domain.bakery.controller.req.BakeryCreateRequest;
+import com.breaditnow.owner.domain.bakery.controller.req.BakeryUpdateRequest;
+import com.breaditnow.owner.domain.bakery.controller.res.BakeryResponse;
+import com.breaditnow.owner.domain.bakery.service.BakeryService;
 
 @SpringBootTest
 @Transactional
